# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Gets short link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short link to get",
                        "name": "ShortLink",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteShortLinkRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/create_link": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "summary": "Gets create link page HTML",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Creates a link",
                "parameters": [
                    {
                        "description": "Short URL",
                        "name": "short_url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Long URL",
                        "name": "long_url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShortLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/delete_link": {
            "delete": {
                "summary": "Tries to delete the short link",
                "parameters": [
                    {
                        "description": "Short link to delete",
                        "name": "ShortLink",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteShortLinkRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/get_links": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Gets user's short links",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit of a number of user's short links",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum amount of links to show",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetUserShortLinksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/get_subscriptions": {
            "get": {
                "summary": "Gets all subscriptions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetSubscriptionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "summary": "Gets login page HTML",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Logins a user",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.redirectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "summary": "Logs out a user",
                "responses": {
                    "307": {
                        "description": "Temporary Redirect",
                        "schema": {
                            "$ref": "#/definitions/handlers.redirectResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/register": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "summary": "Gets register page HTML",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Registers a user",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.redirectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/subscriptions": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "summary": "Gets subscription page HTML",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "summary": "Gets user from session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.CreateShortLinkResponse": {
            "type": "object",
            "properties": {
                "link": {
                    "$ref": "#/definitions/postgresDB.Link"
                }
            }
        },
        "handlers.DeleteShortLinkRequest": {
            "type": "object",
            "properties": {
                "short_link": {
                    "type": "string"
                }
            }
        },
        "handlers.GetSubscriptionsResponse": {
            "type": "object",
            "properties": {
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgresDB.Subscription"
                    }
                }
            }
        },
        "handlers.GetUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/postgresDB.User"
                }
            }
        },
        "handlers.GetUserShortLinksResponse": {
            "type": "object",
            "properties": {
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgresDB.Link"
                    }
                },
                "user": {
                    "$ref": "#/definitions/postgresDB.User"
                }
            }
        },
        "handlers.redirectResponse": {
            "type": "object"
        },
        "postgresDB.Link": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "longUrl": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                },
                "userEmail": {
                    "type": "string"
                }
            }
        },
        "postgresDB.Subscription": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "totalUrls": {
                    "type": "integer"
                }
            }
        },
        "postgresDB.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "passwordHash": {
                    "type": "string"
                },
                "urlsLeft": {
                    "type": "integer"
                }
            }
        }
    }
}